Moved fetch meetings to api calls helper
